+++ out/glsl/format.provider.js
@@ -65,6 +65,8 @@ class GlslFormatProvider {
                     */
                     // remove extra new line
                     s = s.replace(/\n\s*\n\s*\n/g, '\n\n');
+                    // #pragma
+                    s = s.replace(/[^\S\r\n]*#pragma(.+?)\s*:\s*(\S+)[^\S\r\n]*/g, '#pragma$1:$2');
                 }
                 return s;
             });