+++ src/glsl/options.ts
@@ -17,11 +17,16 @@ export default class GlslOptions {
 	public vertex: string;
 	public uniforms: object;
 	public textures: object;
-	public extensions: string[];
-	public doubleSided: boolean;
 	public timeout: number;
+	public extensions: string[];
+	public antialias: boolean;
+	public doubleSided: boolean;
 	public refreshOnChange: boolean;
 	public refreshOnSave: boolean;
+	public recordMethod: string;
+	public recordDuration: number;
+	public recordWidth: number;
+	public recordHeight: number;
 	public mode?: 'mesh' | 'torus' | 'spere' | 'box' | 'flat';
 
 	constructor(webview: vscode.Webview = null, extensionPath: string = null) {
@@ -38,23 +43,33 @@ export default class GlslOptions {
 		this.uniforms = config['uniforms'] || {};
 		this.timeout = config['timeout'] || 0;
 		this.extensions = config['extensions'] || [];
+		this.antialias = config['antialias'] || false;
 		this.doubleSided = config['doubleSided'] || false;
 		this.refreshOnChange = config['refreshOnChange'] || false;
 		this.refreshOnSave = config['refreshOnSave'] || false;
+		this.recordMethod = config['recordMethod'] || 'MediaRecorder';
+		// tslint:disable-next-line: triple-equals
+		this.recordDuration = config['recordDuration'] != null ? config['recordDuration'] : 10;
+		// tslint:disable-next-line: triple-equals
+		this.recordWidth = config['recordWidth'] != null ? config['recordWidth'] : 1024;
+		// tslint:disable-next-line: triple-equals
+		this.recordHeight = config['recordHeight'] != null ? config['recordHeight'] : 1024;
 		if (webview !== null && extensionPath !== null) {
 			const workpath = this.getWorkpath_(webview, extensionPath, workspaceFolder, uri);
 			const folder = this.getFolder_(extensionPath, workspaceFolder, uri);
 			const resources = this.getResources_(webview, extensionPath);
-			if (workspaceFolder) {
-				Object.keys(textures).forEach(key => {
-					let texture = textures[key];
-					if (texture.indexOf('http') === -1) {
+			Object.keys(textures).forEach(key => {
+				let texture = textures[key];
+				if (texture.indexOf('http') === -1) {
+					if (workspaceFolder) {
 						texture = path.join(workspaceFolder, texture);
 						texture = path.relative(folder, texture);
+					} else {
+						texture = path.join(folder, texture);
 					}
-					textures[key] = texture;
-				});
-			}
+				}
+				textures[key] = texture;
+			});
 			this.workpath = workpath;
 			this.folder = folder;
 			this.resources = resources;